+++ data-raw/frac_style.R
@@ -1,0 +1,27 @@
+numerators <- c(
+  "\u2070",
+  "\u00B9",
+  "\u00B2",
+  "\u00B3",
+  "\u2074",
+  "\u2075",
+  "\u2076",
+  "\u2077",
+  "\u2078",
+  "\u2079"
+)
+
+denominators <- c(
+  "\u2080",
+  "\u2081",
+  "\u2082",
+  "\u2083",
+  "\u2084",
+  "\u2085",
+  "\u2086",
+  "\u2087",
+  "\u2088",
+  "\u2089"
+)
+
+usethis::use_data(numerators, denominators, internal = TRUE, overwrite = TRUE)