+++ examples/frac_style.R
@@ -1,0 +1,2 @@
+frac_style(fracture(0.5))
+frac_style(fracture(c(0.5, 1.5), mixed = TRUE))