+++ NEWS.md
@@ -1,5 +1,7 @@
 # fracture (development version)
 
+* Fixed bug with fracture math where only the first element would be returned.
+* Add `frac_style()` which formats fractures using Unicode superscripts and subscripts.
 * Add `is.frac_mat()`.
 
 # fracture 0.1.0